[#40] move files videolist from main to common to use as component in profile page
